Project refactored packages and LogBack added
