+++ src/main/java/com/kainattu/portal/service/PortalServiceApplication.java
@@ -1,13 +1,17 @@
-package com.springBoot;
+package com.kainattu.portal.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * @author Palanive Muthu Gounder
+ *
+ */
 @SpringBootApplication
-public class SpringJwtSecurityApplication {
+public class PortalServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringJwtSecurityApplication.class, args);
+		SpringApplication.run(PortalServiceApplication.class, args);
 	}
 
 }
