+++ src/main/java/com/kainattu/portal/service/config/auth/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,5 @@
-package com.springBoot.config;
+package com.kainattu.portal.service.config.auth;
+
 import java.io.IOException;
 import java.io.Serializable;
 import javax.servlet.http.HttpServletRequest;
@@ -6,9 +7,12 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
+
 @Component
 public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint, Serializable {
+	
 	private static final long serialVersionUID = -7858869558953243875L;
+
 	@Override
 	public void commence(HttpServletRequest request, HttpServletResponse response,
 			AuthenticationException authException) throws IOException {
