+++ src/main/java/com/kainattu/portal/service/config/auth/WebSecurityConfig.java
@@ -1,5 +1,7 @@
-package com.springBoot.config;
+package com.kainattu.portal.service.config.auth;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,13 +21,16 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-	
+
+	private final Logger logger = LoggerFactory.getLogger(getClass());
+
 	@Autowired
 	private JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
 	@Autowired
 	private UserDetailsService jwtUserDetailsService;
 	@Autowired
 	private JwtRequestFilter jwtRequestFilter;
+
 	@Autowired
 	public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
 		// configure AuthenticationManager so that it knows from where to load
@@ -33,30 +38,31 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 		// Use BCryptPasswordEncoder
 		auth.userDetailsService(jwtUserDetailsService).passwordEncoder(passwordEncoder());
 	}
+
 	@Bean
 	public PasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
 	}
+
 	@Bean
 	@Override
 	public AuthenticationManager authenticationManagerBean() throws Exception {
 		return super.authenticationManagerBean();
 	}
+
 	@Override
 	protected void configure(HttpSecurity httpSecurity) throws Exception {
 		// We don't need CSRF for this example
 		httpSecurity.csrf().disable()
-		// dont authenticate this particular request
-		.authorizeRequests()
-		.antMatchers("/authenticate","/register","/h2-console").permitAll()
-		.antMatchers("/admin/**").hasAuthority("Admin")
-		.antMatchers("/user/**").hasAuthority("User").
-		// all other requests need to be authenticated
-		anyRequest().authenticated().and().
-		// make sure we use stateless session; session won't be used to
-		// store user's state.
-		exceptionHandling().authenticationEntryPoint(jwtAuthenticationEntryPoint).and().sessionManagement()
-		.sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+				// dont authenticate this particular request
+				.authorizeRequests().antMatchers("/authenticate", "/register", "/h2-console").permitAll()
+				.antMatchers("/admin/**").hasAuthority("Admin").antMatchers("/user/**").hasAuthority("User").
+				// all other requests need to be authenticated
+				anyRequest().authenticated().and().
+				// make sure we use stateless session; session won't be used to
+				// store user's state.
+				exceptionHandling().authenticationEntryPoint(jwtAuthenticationEntryPoint).and().sessionManagement()
+				.sessionCreationPolicy(SessionCreationPolicy.STATELESS);
 		// Add a filter to validate the tokens with every request
 		httpSecurity.addFilterBefore(jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
 	}
