+++ src/main/java/com/kainattu/portal/service/model/auth/DAOUser.java
@@ -1,4 +1,4 @@
-package com.springBoot.model;
+package com.kainattu.portal.service.model.auth;
 import java.util.HashSet;
 import java.util.Set;
