+++ src/main/java/com/kainattu/portal/service/repo/auth/UserDao.java
@@ -1,9 +1,9 @@
-package com.springBoot.repo;
+package com.kainattu.portal.service.repo.auth;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.springBoot.model.DAOUser;
+import com.kainattu.portal.service.model.auth.DAOUser;
 
 @Repository
 public interface UserDao extends CrudRepository<DAOUser, Integer> {
