+++ src/main/resources/application-cloud2.properties
@@ -1,0 +1,8 @@
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.naming_strategy=org.hibernate.cfg.EJB3NamingStrategy
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+
+spring.h2.console.enabled=true
