+++ src/main/resources/application-local.properties
@@ -1,0 +1,10 @@
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.naming_strategy=org.hibernate.cfg.EJB3NamingStrategy
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+
+#Add DB connection details.
+
+#spring.h2.console.enabled=true
