+++ src/main/resources/application.properties
@@ -1,4 +1,2 @@
-spring.h2.console.enabled=true
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-jwt.secret=javainuse
+spring.profiles.active=@activatedProperties@
+jwt.secret=kainattu
