+++ src/main/resources/logback.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+
+	<property name="LOG_PATH" value="${log.file.location:-.}" />
+	<property name="LOG_ARCHIVE" value="${LOG_PATH}/archive" />
+	<property name="PATTERN"
+		value="%d{HH:mm:ss.SSS} [%thread] %-5level %logger{35} - %msg%n" />
+
+	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+		<layout class="ch.qos.logback.classic.PatternLayout">
+			<Pattern>${PATTERN}</Pattern>
+		</layout>
+	</appender>
+
+	<appender name="FILE"
+		class="ch.qos.logback.core.rolling.RollingFileAppender">
+		<file>${LOG_PATH}/appLog.log</file>
+		<rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+			<!-- daily rollover -->
+			<fileNamePattern>${LOG_ARCHIVE}/appLog.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
+			<timeBasedFileNamingAndTriggeringPolicy
+				class="ch.qos.logback.core.rolling.SizeAndTimeBasedFNATP">
+				<!-- or whenever the file size reaches 50MB -->
+				<maxFileSize>200MB</maxFileSize>
+			</timeBasedFileNamingAndTriggeringPolicy>
+			<!-- keep 30 days' worth of history -->
+			<maxHistory>30</maxHistory>
+		</rollingPolicy>
+		<encoder >
+			<charset>UTF-8</charset>
+				<Pattern>${PATTERN}</Pattern>
+		</encoder>
+	</appender>
+
+	<logger name="com.kainattu.portal" level="DEBUG" additivity="false">
+	<appender-ref ref="STDOUT" />
+	<appender-ref ref="FILE" />
+	</logger>
+
+
+	<root level="INFO" additivity="false">
+		<appender-ref ref="STDOUT" />
+		<appender-ref ref="FILE" />
+	</root>
+</configuration>
